drm/nv50: fix context control on NVAx chipsets
Yes, I'm aware that if you look at a mmio-trace you'll see what looks like
NVIDIA writing 0x00000002 here.  I believe this to be a bug in mmio-trace.
If you run renouveau and hunt down the context immediately after it's
created, and before any accel is done, you'll see that 0x00000000 is the
value that's really written.

Now, the interesting thing here is that the ctxprogs we have for NVAx
chipsets actually have instructions in them that cause them to skip all the
context swapping instructions if 0x00000002 is found there...

This isn't the case on the chipsets that currently work with nouveau, it'll
load/save the value as it does with the other per-channel state, but to my
understanding does not change its behaviour as a result.

What conclusions you draw from this, I'll let you decide...