drm/nouveau: don't disable IRQs when we get an unhandled PFIFO intr
This was probably to avoid getting unending floods of IRQs we don't know
how to handle.  However, we should probably just discover those situations
and learn how to deal with them instead.  Spent far too much time trying
to figure out why PGRAPH was stalling with PE and not sending an IRQ so
we could deal with it.